IUI-41 Display a default message when there's no data in the table.
Originally I was going to put the ng-if on div that wraps the table so
that non of it would display but then I got to thinking that might be
awkward for the user. The error message itself assumes the knowledge of
the existence of a table. If the user didn’t know there was a table
there, they wouldn’t understand the message at all.

By putting the message inside of the table, the user understands that
there’s a table that is supposed to be there. They even get to see what
the columns would have been. The `colspan` set to the length of the
column heads allows for the message to run across the columns.